Rename YuTable to DataTable
